Moved functional block O display message to dedicated class
