+++ lib/CMakeLists.txt
@@ -98,6 +98,7 @@ if(LIBOCPP_ENABLE_V201)
             ocpp/v201/functional_blocks/security.cpp
             ocpp/v201/functional_blocks/data_transfer.cpp
             ocpp/v201/functional_blocks/reservation.cpp
+            ocpp/v201/functional_blocks/display_message.cpp
     )
     add_subdirectory(ocpp/v201/messages)
 endif()
