+++ lib/ocpp/v201/charge_point.cpp
@@ -9,7 +9,6 @@
 #include <ocpp/v201/message_dispatcher.hpp>
 #include <ocpp/v201/messages/FirmwareStatusNotification.hpp>
 #include <ocpp/v201/messages/LogStatusNotification.hpp>
-#include <ocpp/v201/messages/NotifyDisplayMessages.hpp>
 #include <ocpp/v201/notify_report_requests_splitter.hpp>
 
 #include <optional>
@@ -29,10 +28,6 @@ namespace v201 {
 const auto DEFAULT_BOOT_NOTIFICATION_RETRY_INTERVAL = std::chrono::seconds(30);
 const auto DEFAULT_MESSAGE_QUEUE_SIZE_THRESHOLD = 2E5;
 const auto DEFAULT_MAX_MESSAGE_SIZE = 65000;
-
-static DisplayMessageContent message_content_to_display_message_content(const MessageContent& message_content);
-static std::optional<MessageInfo> display_message_to_message_info_type(const DisplayMessage& display_message);
-static DisplayMessage message_info_to_display_message(const MessageInfo& message_info);
 
 ChargePoint::ChargePoint(const std::map<int32_t, int32_t>& evse_connector_structure,
                          std::shared_ptr<DeviceModel> device_model, std::shared_ptr<DatabaseHandler> database_handler,
@@ -944,6 +939,14 @@ void ChargePoint::initialize(const std::map<int32_t, int32_t>& evse_connector_st
                                                           *this->connectivity_manager.get(),
                                                           *this->database_handler.get(), *this->evse_security.get());
     this->authorization->start_auth_cache_cleanup_thread();
+    
+    if (device_model->get_optional_value<bool>(ControllerComponentVariables::DisplayMessageCtrlrAvailable)
+            .value_or(false)) {
+        this->display_message = std::make_unique<DisplayMessageBlock>(
+            *this->message_dispatcher, *this->device_model, *this->evse_manager,
+            this->callbacks.get_display_message_callback.value(), this->callbacks.set_display_message_callback.value(),
+            this->callbacks.clear_display_message_callback.value());
+    }
 
     if (this->callbacks.configure_network_connection_profile_callback.has_value()) {
         this->connectivity_manager->set_configure_network_connection_profile_callback(
@@ -1094,13 +1097,13 @@ void ChargePoint::handle_message(const EnhancedMessage<v201::MessageType>& messa
             this->handle_clear_variable_monitoring_req(json_message);
             break;
         case MessageType::GetDisplayMessages:
-            this->handle_get_display_message(json_message);
+            this->display_message->handle_message(message);
             break;
         case MessageType::SetDisplayMessage:
-            this->handle_set_display_message(json_message);
+            this->display_message->handle_message(message);
             break;
         case MessageType::ClearDisplayMessage:
-            this->handle_clear_display_message(json_message);
+            this->display_message->handle_message(message);
             break;
         case MessageType::CostUpdated:
             this->handle_costupdated_req(json_message);
@@ -1591,18 +1594,6 @@ ChargePoint::set_variables_internal(const std::vector<SetVariableData>& set_vari
     }
 
     return response;
-}
-
-std::optional<int32_t> ChargePoint::get_transaction_evseid(const CiString<36>& transaction_id) {
-    for (auto& evse : *this->evse_manager) {
-        if (evse.has_active_transaction()) {
-            if (transaction_id == evse.get_transaction()->get_transaction().transactionId) {
-                return evse.get_id();
-            }
-        }
-    }
-
-    return std::nullopt;
 }
 
 ocpp::ReservationCheckStatus
@@ -2286,7 +2277,7 @@ void ChargePoint::handle_get_transaction_status(const Call<GetTransactionStatusR
     response.messagesInQueue = false;
 
     if (msg.transactionId.has_value()) {
-        if (this->get_transaction_evseid(msg.transactionId.value()).has_value()) {
+        if (this->evse_manager->get_transaction_evseid(msg.transactionId.value()).has_value()) {
             response.ongoingIndicator = true;
         } else {
             response.ongoingIndicator = false;
@@ -2621,7 +2612,7 @@ void ChargePoint::handle_remote_stop_transaction_request(Call<RequestStopTransac
     const auto msg = call.msg;
 
     RequestStopTransactionResponse response;
-    std::optional<int32_t> evseid = get_transaction_evseid(msg.transactionId);
+    std::optional<int32_t> evseid = this->evse_manager->get_transaction_evseid(msg.transactionId);
 
     if (evseid.has_value()) {
         // F03.FR.07: send 'accepted' if there was an ongoing transaction with the given transaction id
@@ -2757,7 +2748,8 @@ void ChargePoint::handle_costupdated_req(const Call<CostUpdatedRequest> call) {
     running_cost.cost = static_cast<double>(call.msg.totalCost);
     running_cost.transaction_id = call.msg.transactionId;
 
-    std::optional<int32_t> transaction_evse_id = get_transaction_evseid(running_cost.transaction_id);
+    std::optional<int32_t> transaction_evse_id =
+        this->evse_manager->get_transaction_evseid(running_cost.transaction_id);
     if (!transaction_evse_id.has_value()) {
         // We just put an error in the log as the spec does not define what to do here. It is not possible to return
         // a 'Rejected' or something in that manner.
@@ -2780,7 +2772,7 @@ void ChargePoint::handle_costupdated_req(const Call<CostUpdatedRequest> call) {
         return;
     }
 
-    const std::optional<int32_t> evse_id_opt = get_transaction_evseid(running_cost.transaction_id);
+    const std::optional<int32_t> evse_id_opt = this->evse_manager->get_transaction_evseid(running_cost.transaction_id);
     if (!evse_id_opt.has_value()) {
         EVLOG_warning << "Can not set running cost triggers as there is no evse id found with the transaction id from "
                          "the incoming CostUpdatedRequest";
@@ -3324,143 +3316,6 @@ void ChargePoint::handle_clear_variable_monitoring_req(Call<ClearVariableMonitor
     this->message_dispatcher->dispatch_call_result(call_result);
 }
 
-void ChargePoint::handle_get_display_message(const Call<GetDisplayMessagesRequest> call) {
-    GetDisplayMessagesResponse response;
-    if (!this->callbacks.get_display_message_callback.has_value()) {
-        response.status = GetDisplayMessagesStatusEnum::Unknown;
-        ocpp::CallResult<GetDisplayMessagesResponse> call_result(response, call.uniqueId);
-        this->message_dispatcher->dispatch_call_result(call_result);
-        return;
-    }
-
-    // Call 'get display message callback' to get all display messages from the charging station.
-    const std::vector<DisplayMessage> display_messages = this->callbacks.get_display_message_callback.value()(call.msg);
-
-    NotifyDisplayMessagesRequest messages_request;
-    messages_request.requestId = call.msg.requestId;
-    messages_request.messageInfo = std::vector<MessageInfo>();
-    // Convert all display messages from the charging station to the correct format. They will not be included if
-    // they do not have the required values. That's why we wait with sending the response until we converted all
-    // display messages, because we then know if there are any.
-    for (const auto& display_message : display_messages) {
-        const std::optional<MessageInfo> message_info = display_message_to_message_info_type(display_message);
-        if (message_info.has_value()) {
-            messages_request.messageInfo->push_back(message_info.value());
-        }
-    }
-
-    // Send 'accepted' back to the CSMS if there is at least one message and send all the messages in another
-    // request.
-    if (messages_request.messageInfo.value().empty()) {
-        response.status = GetDisplayMessagesStatusEnum::Unknown;
-        ocpp::CallResult<GetDisplayMessagesResponse> call_result(response, call.uniqueId);
-        this->message_dispatcher->dispatch_call_result(call_result);
-        return;
-    } else {
-        response.status = GetDisplayMessagesStatusEnum::Accepted;
-        ocpp::CallResult<GetDisplayMessagesResponse> call_result(response, call.uniqueId);
-        this->message_dispatcher->dispatch_call_result(call_result);
-    }
-
-    // Send display messages. The response is empty, so we don't have to get that back.
-    // Sending multiple messages is not supported for now, because there is no need to split them up (yet).
-    ocpp::Call<NotifyDisplayMessagesRequest> request(messages_request);
-    this->message_dispatcher->dispatch_call(request);
-}
-
-void ChargePoint::handle_set_display_message(const Call<SetDisplayMessageRequest> call) {
-    SetDisplayMessageResponse response;
-    if (!this->callbacks.set_display_message_callback.has_value()) {
-        response.status = DisplayMessageStatusEnum::Rejected;
-        ocpp::CallResult<SetDisplayMessageResponse> call_result(response, call.uniqueId);
-        this->message_dispatcher->dispatch_call_result(call_result);
-        return;
-    }
-
-    // Check if display messages are available, priority and message format are supported and if the given
-    // transaction is running, if a transaction id was included in the message.
-    bool error = false;
-    const std::optional<bool> display_message_available =
-        this->device_model->get_optional_value<bool>(ControllerComponentVariables::DisplayMessageCtrlrAvailable);
-    const std::string supported_priorities =
-        this->device_model->get_value<std::string>(ControllerComponentVariables::DisplayMessageSupportedPriorities);
-    const std::string supported_message_formats =
-        this->device_model->get_value<std::string>(ControllerComponentVariables::DisplayMessageSupportedFormats);
-
-    const std::vector<std::string> priorities = split_string(supported_priorities, ',', true);
-    const std::vector<std::string> formats = split_string(supported_message_formats, ',', true);
-    const auto& supported_priority_it = std::find(
-        priorities.begin(), priorities.end(), conversions::message_priority_enum_to_string(call.msg.message.priority));
-    const auto& supported_format_it = std::find(
-        formats.begin(), formats.end(), conversions::message_format_enum_to_string(call.msg.message.message.format));
-
-    // Check if transaction is valid: this is the case if there is no transaction id, or if the transaction id
-    // belongs to a running transaction.
-    const bool transaction_valid = (!call.msg.message.transactionId.has_value() or
-                                    get_transaction_evseid(call.msg.message.transactionId.value()) != std::nullopt);
-
-    // Check if display messages are available.
-    if (!display_message_available.has_value() or !display_message_available.value()) {
-        error = true;
-        response.status = DisplayMessageStatusEnum::Rejected;
-    }
-    // Check if the priority is supported.
-    else if (supported_priority_it == priorities.end()) {
-        error = true;
-        response.status = DisplayMessageStatusEnum::NotSupportedPriority;
-    }
-    // Check if the message format is supported.
-    else if (supported_format_it == formats.end()) {
-        error = true;
-        response.status = DisplayMessageStatusEnum::NotSupportedMessageFormat;
-    }
-    // Check if transaction is valid.
-    else if (!transaction_valid) {
-        error = true;
-        response.status = DisplayMessageStatusEnum::UnknownTransaction;
-    }
-    // Check if message state is supported.
-    else if (call.msg.message.state.has_value()) {
-        const std::optional<std::string> supported_states = this->device_model->get_optional_value<std::string>(
-            ControllerComponentVariables::DisplayMessageSupportedStates);
-        if (supported_states.has_value()) {
-            const std::vector<std::string> states = split_string(supported_states.value(), ',', true);
-            const auto& supported_states_it =
-                std::find(states.begin(), states.end(),
-                          conversions::message_state_enum_to_string(call.msg.message.state.value()));
-            if (supported_states_it == states.end()) {
-                error = true;
-                response.status = DisplayMessageStatusEnum::NotSupportedState;
-            }
-        }
-    }
-
-    if (error) {
-        ocpp::CallResult<SetDisplayMessageResponse> call_result(response, call.uniqueId);
-        this->message_dispatcher->dispatch_call_result(call_result);
-        return;
-    }
-
-    const DisplayMessage message = message_info_to_display_message(call.msg.message);
-    response = this->callbacks.set_display_message_callback.value()({message});
-    ocpp::CallResult<SetDisplayMessageResponse> call_result(response, call.uniqueId);
-    this->message_dispatcher->dispatch_call_result(call_result);
-}
-
-void ChargePoint::handle_clear_display_message(const Call<ClearDisplayMessageRequest> call) {
-    ClearDisplayMessageResponse response;
-    if (!this->callbacks.clear_display_message_callback.has_value()) {
-        EVLOG_error << "Received a clear display message request, but callback is not implemented.";
-        response.status = ClearMessageStatusEnum::Unknown;
-        ocpp::CallResult<ClearDisplayMessageResponse> call_result(response, call.uniqueId);
-        this->message_dispatcher->dispatch_call_result(call_result);
-    }
-
-    response = this->callbacks.clear_display_message_callback.value()(call.msg);
-    ocpp::CallResult<ClearDisplayMessageResponse> call_result(response, call.uniqueId);
-    this->message_dispatcher->dispatch_call_result(call_result);
-}
-
 std::optional<DataTransferResponse> ChargePoint::data_transfer_req(const CiString<255>& vendorId,
                                                                    const std::optional<CiString<50>>& messageId,
                                                                    const std::optional<json>& data) {
@@ -3830,77 +3685,6 @@ void ChargePoint::send_not_implemented_error(const MessageId unique_message_id, 
         const auto call_error = CallError(unique_message_id, "NotImplemented", "", json({}));
         this->message_dispatcher->dispatch_call_error(call_error);
     }
-}
-
-// Static functions
-
-///
-/// \brief Convert message content from OCPP spec to DisplayMessageContent.
-/// \param message_content  The struct to convert.
-/// \return The converted struct.
-///
-static DisplayMessageContent message_content_to_display_message_content(const MessageContent& message_content) {
-    DisplayMessageContent result;
-    result.message = message_content.content;
-    result.message_format = message_content.format;
-    result.language = message_content.language;
-    return result;
-}
-
-///
-/// \brief Convert display message to MessageInfo from OCPP.
-/// \param display_message  The struct to convert.
-/// \return The converted struct.
-///
-static std::optional<MessageInfo> display_message_to_message_info_type(const DisplayMessage& display_message) {
-    // Each display message should have an id and p[riority, this is required for OCPP.
-    if (!display_message.id.has_value()) {
-        EVLOG_error << "Can not convert DisplayMessage to MessageInfo: No id is provided, which is required by OCPP.";
-        return std::nullopt;
-    }
-
-    if (!display_message.priority.has_value()) {
-        EVLOG_error
-            << "Can not convert DisplayMessage to MessageInfo: No priority is provided, which is required by OCPP.";
-        return std::nullopt;
-    }
-
-    MessageInfo info;
-    info.message.content = display_message.message.message;
-    info.message.format =
-        (display_message.message.message_format.has_value() ? display_message.message.message_format.value()
-                                                            : MessageFormatEnum::UTF8);
-    info.message.language = display_message.message.language;
-    info.endDateTime = display_message.timestamp_to;
-    info.startDateTime = display_message.timestamp_from;
-    info.id = display_message.id.value();
-    info.priority = display_message.priority.value();
-    info.state = display_message.state;
-    info.transactionId = display_message.identifier_id;
-
-    // Note: component is (not yet?) supported for display messages in libocpp.
-
-    return info;
-}
-
-///
-/// \brief Convert message info from OCPP to DisplayMessage.
-/// \param message_info The struct to convert.
-/// \return The converted struct.
-///
-static DisplayMessage message_info_to_display_message(const MessageInfo& message_info) {
-    DisplayMessage display_message;
-
-    display_message.id = message_info.id;
-    display_message.priority = message_info.priority;
-    display_message.state = message_info.state;
-    display_message.timestamp_from = message_info.startDateTime;
-    display_message.timestamp_to = message_info.endDateTime;
-    display_message.identifier_id = message_info.transactionId;
-    display_message.identifier_type = IdentifierType::TransactionId;
-    display_message.message = message_content_to_display_message_content(message_info.message);
-
-    return display_message;
 }
 
 } // namespace v201
